+++ extensions/fotmob/src/team.tsx
@@ -1,5 +1,6 @@
-import { Icon, LaunchProps, List, Toast, showToast, useNavigation } from "@raycast/api";
-import TeamView from "./views/team/TeamView";
+import type { LaunchProps } from "@raycast/api";
+import { Icon, List, Toast, showToast, useNavigation } from "@raycast/api";
+import TeamView from "@/views/team/TeamView";
 
 export default function Command(props: LaunchProps<{ arguments: Arguments.Team }>) {
   const { teamId } = props.arguments;
@@ -25,8 +26,6 @@ export default function Command(props: LaunchProps<{ arguments: Arguments.Team }
         <List.EmptyView icon={Icon.XMarkCircleFilled} description="Team ID must be a number" />
       </List>
     );
-
-    return null;
   }
 
   return <TeamView id={teamId} />;